third_party/blink/web_tests/fast/xpath/4XPath/Borrowed/sr_20021217-expected.txt

PASS node_names is "bcd"
PASS node_names is "b"
PASS successfullyParsed is true

TEST COMPLETE
